[LLHD] Change simulator time granularity to ps
* Previously only the time unit was ignored, and time delays were assumed to be in `ns`. This makes the base unit for the simulator be `ps` and considers the time unit when lowering.
* Promote the simulator's state time values to uint64_t to accomodate the new base time unit.

+++ lib/Dialect/LLHD/Simulator/State.h
@@ -24,7 +24,7 @@ struct Time {
   Time() = default;
 
   /// Construct with given time values.
-  Time(unsigned time, unsigned delta, unsigned eps)
+  Time(uint64_t time, uint64_t delta, uint64_t eps)
       : time(time), delta(delta), eps(eps) {}
 
   /// Compare the time values in order of time, delta, eps.
@@ -42,9 +42,9 @@ struct Time {
   /// Get the stored time in a printable format.
   std::string dump();
 
-  unsigned time;
-  unsigned delta;
-  unsigned eps;
+  uint64_t time;
+  uint64_t delta;
+  uint64_t eps;
 
 private:
 };
